quick commit

+++ HI_FAST/test_light.py
@@ -99,9 +99,9 @@ def plot_coldgas(z):
 
     for i in range(len(model_names)):
         index = model_names[i]
-        c = numpy.empty(nGals[index],dtype=numpy.float32)
-        pos =  numpy.asfortranarray(gal[index]['Pos'])
-        mymodule.blas_3dvsdot(c_int(nGals[index]),pos.ctypes.data_as(POINTER(c_float)),c.ctypes.data_as(POINTER(c_float)))
+        c = numpy.empty(nGals[index],dtype=numpy.float64)
+        pos =  numpy.asfortranarray(gal[index]['Pos'].astype(numpy.float64))
+        mymodule.blas_3dvsdot(c_int(nGals[index]),pos.ctypes.data_as(POINTER(c_double)),c.ctypes.data_as(POINTER(c_double)))
         print c
